+++ src/traits/computable_grid.rs
@@ -25,21 +25,21 @@ pub struct ComputeGrid {
 
 #[derive(Debug, PartialEq, Eq, Hash, Default, Clone, Copy)]
 pub struct ComputeGridNode {
-	pub x: usize,
-	pub y: usize,
+	pub x: i32,
+	pub y: i32,
 }
 
 impl ComputeGridNode {
 	pub const ZERO: Self = Self::from_translation(Vec2::ZERO);
 
-	pub const fn new(x: usize, y: usize) -> Self {
+	pub const fn new(x: i32, y: i32) -> Self {
 		Self { x, y }
 	}
 
 	pub const fn from_translation(Vec2 { x, y }: Vec2) -> Self {
 		Self {
-			x: x as usize,
-			y: y as usize,
+			x: x as i32,
+			y: y as i32,
 		}
 	}
 }
